feat #58 person id int type 변경

+++ data/src/main/java/com/dcs/data/remote/datasource/PersonRemoteDataSourceImpl.kt
@@ -25,7 +25,7 @@ class PersonRemoteDataSourceImpl
                     it.data as GetPopularPeopleResponse
                 }
 
-        override suspend fun getPersonDetail(personId: Long): Result<GetPersonDetailResponse> =
+        override suspend fun getPersonDetail(personId: Int): Result<GetPersonDetailResponse> =
             personService
                 .getPersonDetail(
                     personId = personId,
